update service to use Set Favorite Button and functional components
